cleanup heatmap code
