+++ mettascope/src/heatmap.ts
@@ -1,4 +1,4 @@
-import { state } from './common.js'
+import { state, HEATMAP_MIN_OPACITY, HEATMAP_MAX_OPACITY } from './common.js'
 import { getAttr } from './replay.js'
 
 /**
@@ -98,5 +98,61 @@ export class Heatmap {
     getHeat(step: number, x: number, y: number): number {
     this.assertValidPosition(step, x, y)
     return this.data[step][x][y]
+  }
+}
+
+/** Maps a normalized heat value (0-1) to a color using a blue-yellow-red thermal gradient. */
+export function getHeatmapColor(normalizedHeat: number): [number, number, number, number] {
+  // Use square root scaling for better visual distribution.
+  const scaledHeat = Math.sqrt(normalizedHeat)
+  const opacity = (HEATMAP_MIN_OPACITY * 0.5) + (scaledHeat * (HEATMAP_MAX_OPACITY * 0.6))
+  let r: number, g: number, b: number
+
+  if (scaledHeat < 0.5) {
+    // Blue to yellow transition: blue → cyan → green → yellow.
+    const t = scaledHeat * 2  // Normalize to 0-1 range for this half.
+    r = Math.max((t - 0.5) * 1.6, 0)  // Only add red in second half (0.5-1.0) to avoid purple.
+    g = t * 0.8  // Add green from 0 to 0.8 for cyan/green.
+    b = (1 - t) * 0.7  // Fade out blue linearly from 0.7 to 0.
+  } else {
+    // Yellow to red transition preserves red and reduces green.
+    const t = (scaledHeat - 0.5) * 2  // Normalize to 0-1 range for this half.
+    r = 0.8 + (t * 0.2)  // Increase red slightly from 0.8 to 1.0.
+    g = 0.8 * (1 - t)  // Fade out green from 0.8 to 0 for pure red.
+    b = 0  // No blue component in yellow-red range.
+  }
+
+  return [r, g, b, opacity]
+}
+
+/** Gets the maximum heat value for a given step. */
+export function getMaxHeat(step: number): number {
+  let maxHeat = 0
+  for (let x = 0; x < state.replay.map_size[0]; x++) {
+    for (let y = 0; y < state.replay.map_size[1]; y++) {
+      const heat = state.heatmap.getHeat(step, x, y)
+      maxHeat = Math.max(maxHeat, heat)
+    }
+  }
+  return maxHeat
+}
+
+/** Renders heatmap tiles using a provided drawing function. */
+export function renderHeatmapTiles(
+  step: number,
+  drawTile: (x: number, y: number, color: [number, number, number, number]) => void
+): void {
+  const maxHeat = getMaxHeat(step)
+  if (maxHeat === 0) return
+
+  for (let x = 0; x < state.replay.map_size[0]; x++) {
+    for (let y = 0; y < state.replay.map_size[1]; y++) {
+      const heat = state.heatmap.getHeat(step, x, y)
+      if (heat > 0) {
+        const normalizedHeat = heat / maxHeat
+        const color = getHeatmapColor(normalizedHeat)
+        drawTile(x, y, color)
+      }
+    }
   }
 }
