+++ mettascope/src/minimap.ts
@@ -1,5 +1,6 @@
 import * as Common from './common.js'
 import { ctx, state, ui } from './common.js'
+import { renderHeatmapTiles } from './heatmap.js'
 import { parseHtmlColor } from './htmlutils.js'
 import type { PanelInfo } from './panels.js'
 import { getAttr } from './replay.js'
@@ -19,6 +20,13 @@ export function renderMinimapObjects(offset: Vec2f) {
     state.replay.map_size[1],
     parseHtmlColor('#E7D4B7')
   )
+
+  // Draw the heatmap background if enabled.
+  if (state.showHeatmap) {
+    renderHeatmapTiles(state.step, (x: number, y: number, color: [number, number, number, number]) => {
+      ctx.drawSolidRect(x + offset.x(), y + offset.y(), 1, 1, color)
+    })
+  }
 
   // Draw the grid objects on the minimap.
   for (const gridObject of state.replay.grid_objects) {
