+++ mettascope/src/worldmap.ts
@@ -1,6 +1,7 @@
 import * as Common from './common.js'
 import { ctx, setFollowSelection, state, ui, HEATMAP_MIN_OPACITY, HEATMAP_MAX_OPACITY } from './common.js'
 import { Grid } from './grid.js'
+import { renderHeatmapTiles } from './heatmap.js'
 import { type HoverBubble, updateHoverBubble, updateReadout } from './hoverbubbles.js'
 import { parseHtmlColor } from './htmlutils.js'
 import { updateSelection } from './main.js'
@@ -642,62 +643,20 @@ function drawGrid() {
   }
 }
 
-/** Maps a normalized heat value (0-1) to a color using a blue-yellow-red thermal gradient. */
-function getHeatmapColor(normalizedHeat: number): [number, number, number, number] {
-  // Use square root scaling for better visual distribution.
-  const scaledHeat = Math.sqrt(normalizedHeat)
-  const opacity = (HEATMAP_MIN_OPACITY * 0.5) + (scaledHeat * (HEATMAP_MAX_OPACITY * 0.6))
-  let r: number, g: number, b: number
 
-  if (scaledHeat < 0.5) {
-    // Blue to yellow transition: blue → cyan → green → yellow.
-    const t = scaledHeat * 2  // Normalize to 0-1 range for this half.
-    r = Math.max((t - 0.5) * 1.6, 0)  // Only add red in second half (0.5-1.0) to avoid purple.
-    g = t * 0.8  // Add green from 0 to 0.8 for cyan/green.
-    b = (1 - t) * 0.7  // Fade out blue linearly from 0.7 to 0.
-  } else {
-    // Yellow to red transition preserves red and reduces green.
-    const t = (scaledHeat - 0.5) * 2  // Normalize to 0-1 range for this half.
-    r = 0.8 + (t * 0.2)  // Increase red slightly from 0.8 to 1.0.
-    g = 0.8 * (1 - t)  // Fade out green from 0.8 to 0 for pure red.
-    b = 0  // No blue component in yellow-red range.
-  }
-
-  return [r, g, b, opacity]
-}
 
 /** Draws the heatmap overlay. */
 function drawHeatmap() {
   if (state.showHeatmap) {
-    // Find the maximum heat value for normalization
-    let maxHeat = 0
-    for (let x = 0; x < state.replay.map_size[0]; x++) {
-      for (let y = 0; y < state.replay.map_size[1]; y++) {
-        const heat = state.heatmap.getHeat(state.step, x, y)
-        maxHeat = Math.max(maxHeat, heat)
-      }
-    }
-
-    if (maxHeat === 0) return
-
-    for (let x = 0; x < state.replay.map_size[0]; x++) {
-      for (let y = 0; y < state.replay.map_size[1]; y++) {
-        const heat = state.heatmap.getHeat(state.step, x, y)
-        if (heat > 0) {
-          const normalizedHeat = heat / maxHeat
-          const color = getHeatmapColor(normalizedHeat)
-
-          // Draw the heatmap overlay with gradient colors
-          ctx.drawSolidRect(
-            x * Common.TILE_SIZE - Common.TILE_SIZE / 2,
-            y * Common.TILE_SIZE - Common.TILE_SIZE / 2,
-            Common.TILE_SIZE,
-            Common.TILE_SIZE,
-            color
-          )
-        }
-      }
-    }
+    renderHeatmapTiles(state.step, (x: number, y: number, color: [number, number, number, number]) => {
+      ctx.drawSolidRect(
+        x * Common.TILE_SIZE - Common.TILE_SIZE / 2,
+        y * Common.TILE_SIZE - Common.TILE_SIZE / 2,
+        Common.TILE_SIZE,
+        Common.TILE_SIZE,
+        color
+      )
+    })
   }
 }
